docs(upgrade): switch file naming scheme, update dependencies, remove unnecessary providers.
Switch to the component.type.ts naming scheme in the PhoneCat tutorial.
Update Angular 2 to beta.7 and Angular 1 to 1.5.0.
Remove the unneeded ROUTER_DIRECTIVES entry from bootstrap providers.

Closes #878
