+++ public/docs/_examples/upgrade/ts/adapter/package.json
@@ -7,18 +7,20 @@
   "author": "",
   "license": "ISC",
   "devDependencies": {
-    "typescript": "1.7.5"
+    "typescript": "1.8.2",
+    "typings": "^0.6.8"
   },
   "dependencies": {
-    "angular2": "2.0.0-beta.3",
+    "angular2": "2.0.0-beta.7",
     "es6-promise": "3.0.2",
     "es6-shim": "0.33.13",
     "reflect-metadata": "0.1.2",
-    "rxjs": "5.0.0-beta.0",
-    "systemjs": "0.19.9",
-    "zone.js": "0.5.11"
+    "rxjs": "5.0.0-beta.2",
+    "systemjs": "0.19.22",
+    "zone.js": "0.5.15"
   },
   "scripts": {
-    "tsc": "tsc -p . -w"
+    "tsc": "tsc -p . -w",
+    "typings": "typings"
   }
 }
