+++ public/docs/_examples/upgrade/ts/adapter/tsconfig.json
@@ -7,6 +7,8 @@
     "experimentalDecorators": true
   },
   "exclude": [
-    "node_modules"
+    "node_modules",
+    "typings/main.d.ts",
+    "typings/main"
   ]
 }
