+++ public/docs/_examples/upgrade/ts/classes/package.1.json
@@ -6,19 +6,19 @@
   "repository": "https://github.com/angular/angular-phonecat",
   "license": "MIT",
   "dependencies": {
-    "systemjs": "0.19.6"
+    "systemjs": "0.19.22"
   },
   "devDependencies": {
     "karma": "^0.12.16",
     "karma-chrome-launcher": "^0.1.4",
     "karma-firefox-launcher": "^0.1.3",
-    "karma-jasmine": "~0.2.0",
+    "karma-jasmine": "~0.3.7",
     "protractor": "^3.0.0",
     "http-server": "^0.6.1",
     "tmp": "0.0.23",
     "bower": "^1.3.1",
     "shelljs": "^0.2.6",
-    "typescript": "1.7.3",
+    "typescript": "1.8.2",
     "typings": "^0.6.8"
   },
   "scripts": {
