+++ public/docs/_examples/upgrade/ts/ng2_components/app/index.html
@@ -14,7 +14,7 @@
   <script src="bower_components/angular-resource/angular-resource.js"></script>
   <script src="../node_modules/es6-shim/es6-shim.min.js"></script>
   <script src="../node_modules/systemjs/dist/system-polyfills.js"></script>
-  <script src="node_modules/angular2/es6/dev/src/testing/shims_for_IE.js"></script>
+  <script src="../node_modules/angular2/es6/dev/src/testing/shims_for_IE.js"></script>
   <script src="../node_modules/angular2/bundles/angular2-polyfills.js"></script>
   <script src="../node_modules/angular2/bundles/angular2.dev.js"></script>
   <script src="../node_modules/angular2/bundles/upgrade.dev.js"></script>
