+++ public/docs/_examples/upgrade/ts/ng2_components/app/js/core/checkmark.pipe.ts
