+++ public/docs/_examples/upgrade/ts/ng2_components/app/js/core/core.module.ts
@@ -1,5 +1,5 @@
 // #docregion
-import {Phones} from './Phones';
+import {Phones} from './phones.service';
 import upgradeAdapter from './upgrade_adapter';
 
 upgradeAdapter.addProvider(Phones);
