+++ public/docs/_examples/upgrade/ts/ng2_components/app/js/phone_detail/phone_detail.module.ts
@@ -1,5 +1,5 @@
 // #docregion
-import PhoneDetail from './PhoneDetail';
+import PhoneDetail from './phone_detail.component';
 import upgradeAdapter from '../core/upgrade_adapter';
 
 export default angular.module('phonecat.detail', [
