+++ public/docs/_examples/upgrade/ts/ng2_components/test/unit/order_by.pipe.spec.ts
@@ -1,9 +1,9 @@
 // #docregion
 import {describe, beforeEachProviders, it, inject} from 'angular2/testing';
 
-import OrderByPipe from '../../app/js/phone_list/OrderByPipe';
+import OrderByPipe from '../../app/js/phone_list/order_by.pipe';
 
-describe('OrderByPipe', () => {
+describe('OrderByPipe', function() {
 
   let input:any[] = [
     {name: 'Nexus S', snippet: 'The Nexus S Phone', images: []},
