+++ public/docs/_examples/upgrade/ts/ng2_final/app/index.html
@@ -10,14 +10,15 @@
   <!-- #docregion scripts -->
   <!-- IE required polyfill -->
   <script src="../node_modules/es6-shim/es6-shim.min.js"></script>
-  <script src="../node_modules/systemjs/dist/system-polyfills.js"></script>  
-  <script src="../node_modules/angular2/es6/dev/src/testing/shims_for_IE.js"></script>   
+  <script src="../node_modules/systemjs/dist/system-polyfills.js"></script>
+  <script src="../node_modules/angular2/es6/dev/src/testing/shims_for_IE.js"></script>
   <!-- #enddocregion ie-polyfills -->
 
   <script src="../node_modules/angular2/bundles/angular2-polyfills.js"></script>
   <script src="../node_modules/systemjs/dist/system.src.js"></script>
   <script src="../node_modules/rxjs/bundles/Rx.js"></script>
   <script src="../node_modules/angular2/bundles/angular2.dev.js"></script>
+  <script src="../node_modules/angular2/bundles/http.dev.js"></script>
   <!-- #docregion ng2-router -->
   <script src="../node_modules/angular2/bundles/router.dev.js"></script>
   <!-- #enddocregion ng2-router -->
@@ -29,7 +30,7 @@
         }
       }
     });
-    System.import('js/app');
+    System.import('js/main');
   </script>
   <!-- #enddocregion scripts -->
 </head>
