+++ public/docs/_examples/upgrade/ts/ng2_final/app/js/core/checkmark.pipe.ts
