+++ public/docs/_examples/upgrade/ts/ng2_final/app/js/phone_detail/phone_detail.component.ts
@@ -1,21 +1,22 @@
 // #docregion
+// #docregion top
 import {Component, Inject} from 'angular2/core';
-import {Phones, Phone} from '../core/Phones';
-
-interface PhoneRouteParams {
-  phoneId: string
-}
+import {RouteParams} from 'angular2/router';
+import {Phones, Phone} from '../core/phones.service';
+import {CheckmarkPipe} from '../core/checkmark.pipe';
 
 @Component({
   selector: 'pc-phone-detail',
-  templateUrl: 'js/phone_detail/phone_detail.html'
+  templateUrl: 'js/phone_detail/phone_detail.html',
+  pipes: [CheckmarkPipe]
 })
 class PhoneDetail {
+// #enddocregion top
   phone:Phone = undefined;
   mainImageUrl:string;
-  constructor(@Inject('$routeParams') $routeParams:PhoneRouteParams,
+  constructor(params:RouteParams,
               phones:Phones) {
-    phones.get($routeParams.phoneId)
+    phones.get(params.get('phoneId'))
       .subscribe(phone => {
         this.phone = phone;
         this.mainImageUrl = phone.images[0];
