+++ public/docs/_examples/upgrade/ts/ng2_final/app/js/phone_list/order_by.pipe.ts
