+++ public/docs/_examples/upgrade/ts/ng2_final/app/js/phone_list/phone_filter.pipe.ts
@@ -1,6 +1,6 @@
 // #docregion
 import {Pipe} from 'angular2/core';
-import {Phone} from '../core/Phones';
+import {Phone} from '../core/phones.service';
 
 @Pipe({name: 'phoneFilter'})
 export default class PhoneFilterPipe {
