+++ public/docs/_examples/upgrade/ts/ng2_final/app/js/phone_list/phone_list.component.ts
@@ -1,11 +1,17 @@
+// #docregion full
 // #docregion top
 import {Component} from 'angular2/core';
+import {RouterLink} from 'angular2/router';
 import {Observable} from 'rxjs';
-import {Phones, Phone} from '../core/Phones';
+import {Phones, Phone} from '../core/phones.service';
+import PhoneFilterPipe from './phone_filter.pipe';
+import OrderByPipe from './order_by.pipe';
 
 @Component({
   selector: 'pc-phone-list',
-  templateUrl: 'js/phone_list/phone_list.html'
+  templateUrl: 'js/phone_list/phone_list.html',
+  pipes: [PhoneFilterPipe, OrderByPipe],
+  directives: [RouterLink]
 })
 class PhoneList {
 // #enddocregion top
