+++ public/docs/_examples/upgrade/ts/ng2_final/test/unit/phone_filter.pipe.spec.ts
@@ -1,10 +1,10 @@
 // #docregion
 import {describe, beforeEachProviders, it, inject} from 'angular2/testing';
 
-import PhoneFilterPipe from '../../app/js/phone_list/PhoneFilterPipe';
-import {Phone} from '../../app/js/core/Phones';
+import PhoneFilterPipe from '../../app/js/phone_list/phone_filter.pipe';
+import {Phone} from '../../app/js/core/phones.service';
 
-describe('PhoneFilterPipe', function() {
+describe('PhoneFilterPipe', () => {
 
   let phones:Phone[] = [
     {name: 'Nexus S', snippet: 'The Nexus S Phone', images: []},
