+++ public/docs/_examples/upgrade/ts/ng2_final/test/unit/phone_list.component.spec.ts
@@ -1,8 +1,15 @@
 // #docregion
-import {provide} from 'angular2/core';
+import {provide, ApplicationRef} from 'angular2/core';
 import {HTTP_PROVIDERS} from 'angular2/http';
-import {Observable} from 'rxjs';
-import {FromObservable} from 'rxjs/observable/from';
+import {
+  ROUTER_PROVIDERS
+  ROUTER_PRIMARY_COMPONENT,
+  LocationStrategy
+} from 'angular2/router';
+import {MockApplicationRef} from 'angular2/testing';
+import {MockLocationStrategy} from 'angular2/router/testing';
+import {Observable} from 'rxjs/Rx';
+import 'rxjs/add/observable/fromArray';
 import {
   describe,
   beforeEachProviders,
@@ -11,12 +18,13 @@ import {
   expect,
   TestComponentBuilder
 } from 'angular2/testing';
-import PhoneList from '../../app/js/phone_list/PhoneList';
-import {Phones, Phone} from '../../app/js/core/Phones';
+import AppComponent from '../../app/js/app.component';
+import PhoneList from '../../app/js/phone_list/phone_list.component';
+import {Phones, Phone} from '../../app/js/core/phones.service';
 
 class MockPhones extends Phones {
   query():Observable<Phone[]> {
-    return FromObservable.create([
+    return Observable.fromArray([
       [{name: 'Nexus S'}, {name: 'Motorola DROID'}]
     ])
   }
@@ -25,8 +33,12 @@ class MockPhones extends Phones {
 describe('PhoneList', () => {
 
   beforeEachProviders(() => [
-    provide(Phones, {useClass: MockPhones}),
-    HTTP_PROVIDERS
+    HTTP_PROVIDERS,
+    ROUTER_PROVIDERS,
+    provide(ApplicationRef, {useClass: MockApplicationRef}),
+    provide(ROUTER_PRIMARY_COMPONENT, {useValue: AppComponent}),
+    provide(LocationStrategy, {useClass: MockLocationStrategy}),
+    provide(Phones, {useClass: MockPhones})
   ]);
 
 
@@ -38,8 +50,8 @@ describe('PhoneList', () => {
       let compiled = fixture.debugElement.nativeElement;
 
       expect(compiled.querySelectorAll('.phone-listing').length).toBe(2);
-      expect(compiled.querySelector('.phone-listing:nth-child(1)')).toHaveText('Nexus S');
-      expect(compiled.querySelector('.phone-listing:nth-child(2)')).toHaveText('Motorola DROID');
+      expect(compiled.querySelector('.phone-listing:nth-child(1)').textContent).toContain('Nexus S');
+      expect(compiled.querySelector('.phone-listing:nth-child(2)').textContent).toContain('Motorola DROID');
     });
   }));
