+++ public/docs/_examples/upgrade/ts/ng2_final/test/unit/phones.service.spec.ts
@@ -1,9 +1,9 @@
 // #docregion
 import {describe, beforeEachProviders, it, inject} from 'angular2/testing';
 import {HTTP_PROVIDERS} from 'angular2/http';
-import {Phones} from '../../app/js/core/Phones';
+import {Phones} from '../../app/js/core/phones.service';
 
-describe('Phones', function() {
+describe('Phones', () => {
 
   // load providers
   beforeEachProviders(() => [Phones, HTTP_PROVIDERS]);
