+++ public/docs/_examples/upgrade/ts/ng2_initial/app/index.html
@@ -16,7 +16,7 @@
   <!-- #docregion ng2 -->
   <script src="../node_modules/es6-shim/es6-shim.js"></script>
   <script src="../node_modules/es6-promise/dist/es6-promise.js"></script>
-  <script src="node_modules/angular2/es6/dev/src/testing/shims_for_IE.js"></script>
+  <script src="../node_modules/angular2/es6/dev/src/testing/shims_for_IE.js"></script>
   <script src="../node_modules/angular2/bundles/angular2-polyfills.js"></script>
   <script src="../node_modules/angular2/bundles/angular2.dev.js"></script>
   <script src="../node_modules/angular2/bundles/upgrade.dev.js"></script>
