+++ public/docs/_examples/upgrade/ts/ng2_initial/app/js/core/core.module.ts
@@ -1,5 +1,5 @@
 // #docregion
-import {Phones} from './Phones';
+import {Phones} from './phones.service';
 import checkmarkFilter from './checkmark.filter';
 import upgradeAdapter from './upgrade_adapter';
