+++ public/docs/_examples/upgrade/ts/ng2_initial/app/js/phone_list/phone_list.controller.ts
@@ -1,5 +1,5 @@
 // #docregion
-import {Phones, Phone} from '../core/Phones';
+import {Phones, Phone} from '../core/phones.service';
 
 class PhoneListCtrl {
   phones:Phone[];
