+++ public/docs/_examples/upgrade/ts/ng2_initial/package.1.json
@@ -6,25 +6,25 @@
   "repository": "https://github.com/angular/angular-phonecat",
   "license": "MIT",
   "dependencies": {
-    "angular2": "2.0.0-beta.3",
+    "angular2": "2.0.0-beta.7",
     "es6-promise": "^3.0.2",
     "es6-shim": "^0.33.3",
     "reflect-metadata": "0.1.2",
-    "rxjs": "5.0.0-beta.0",
-    "zone.js": "0.5.11",
-    "systemjs": "0.19.6"
+    "rxjs": "5.0.0-beta.2",
+    "zone.js": "0.5.15",
+    "systemjs": "0.19.22"
   },
   "devDependencies": {
     "karma": "^0.12.16",
     "karma-chrome-launcher": "^0.1.4",
     "karma-firefox-launcher": "^0.1.3",
-    "karma-jasmine": "~0.2.0",
+    "karma-jasmine": "~0.3.7",
     "protractor": "^3.0.0",
     "http-server": "^0.6.1",
     "tmp": "0.0.23",
     "bower": "^1.3.1",
     "shelljs": "^0.2.6",
-    "typescript": "1.7.3",
+    "typescript": "1.8.2",
     "typings": "^0.6.8"
   },
   "scripts": {
