+++ public/docs/_examples/upgrade/ts/ng2_initial/test/unit/phone_list.controller.spec.ts
@@ -1,9 +1,9 @@
 // #docregion
-import {Observable} from 'rxjs';
+import {Observable} from 'rxjs/Rx';
 import '../../app/js/phone_list/phone_list.module';
-import {Phones} from '../../app/js/core/Phones';
+import {Phones} from '../../app/js/core/phones.service';
 
-import {FromObservable} from 'rxjs/observable/from';
+import 'rxjs/add/observable/fromArray';
 
 describe('PhoneListCtrl', () => {
   var scope, ctrl, $httpBackend;
@@ -18,7 +18,7 @@ describe('PhoneListCtrl', () => {
   }));
 
   beforeEach(inject(function(phones, $rootScope, $controller) {
-    spyOn(phones, 'query').and.returnValue(FromObservable.create([
+    spyOn(phones, 'query').and.returnValue(Observable.fromArray([
       [{name: 'Nexus S'}, {name: 'Motorola DROID'}]
     ]));
     scope = $rootScope.$new();
