+++ public/docs/_examples/upgrade/ts/ng2_initial/test/unit/phones.service.spec.ts
@@ -1,7 +1,7 @@
 // #docregion
 import {describe, beforeEachProviders, it, inject} from 'angular2/testing';
 import {HTTP_PROVIDERS} from 'angular2/http';
-import {Phones} from '../../app/js/core/Phones';
+import {Phones} from '../../app/js/core/phones.service';
 
 describe('Phones', () => {
