+++ public/docs/_examples/upgrade/ts/typescript-conversion/bower.json
@@ -6,15 +6,15 @@
   "license": "MIT",
   "private": true,
   "dependencies": {
-    "angular": "1.5.0-beta.2",
-    "angular-mocks": "1.5.0-beta.2",
+    "angular": "1.5.0",
+    "angular-mocks": "1.5.0",
     "jquery": "~2.1.1",
     "bootstrap": "~3.1.1",
-    "angular-route": "1.5.0-beta.2",
-    "angular-resource": "1.5.0-beta.2",
-    "angular-animate": "1.5.0-beta.2"
+    "angular-route": "1.5.0",
+    "angular-resource": "1.5.0",
+    "angular-animate": "1.5.0"
   },
   "resolutions": {
-    "angular": "1.5.0-beta.2"
+    "angular": "1.5.0"
   }
 }
