+++ public/docs/_examples/upgrade/ts/typescript-conversion/package.1.json
@@ -6,7 +6,7 @@
   "repository": "https://github.com/angular/angular-phonecat",
   "license": "MIT",
   "dependencies": {
-    "systemjs": "0.19.6"
+    "systemjs": "0.19.22"
   },
   "devDependencies": {
     "bower": "^1.3.1",
@@ -14,11 +14,11 @@
     "karma": "^0.12.16",
     "karma-chrome-launcher": "^0.1.4",
     "karma-firefox-launcher": "^0.1.3",
-    "karma-jasmine": "~0.2.0",
+    "karma-jasmine": "~0.3.7",
     "protractor": "^3.0.0",
     "shelljs": "^0.2.6",
     "tmp": "0.0.23",
-    "typescript": "^1.7.5",
+    "typescript": "^1.8.2",
     "typings": "^0.6.8"
   },
   "scripts": {
